+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/query/plan/cascades/predicates/Placeholder.java
@@ -41,9 +41,9 @@ public class Placeholder extends ValueWithRanges implements WithAlias {
     private final CorrelationIdentifier parameterAlias;
 
     private Placeholder(@Nonnull Value value,
-                       @Nonnull final Set<RangeConstraints> rangeConstraints,
+                       @Nonnull final Set<RangeConstraints> ranges,
                        @Nonnull final CorrelationIdentifier alias) {
-        super(value, rangeConstraints);
+        super(value, ranges);
         this.parameterAlias = alias;
     }
 
@@ -55,6 +55,10 @@ public class Placeholder extends ValueWithRanges implements WithAlias {
     @Nonnull
     public static Placeholder newInstance(@Nonnull Value value, @Nonnull CorrelationIdentifier parameterAlias) {
         return new Placeholder(value, ImmutableSet.of(), parameterAlias);
+    }
+
+    public boolean isConstraining() {
+        return !getRanges().isEmpty();
     }
 
     @Nonnull